migrate update

+++ database/migrations/2025_12_05_114230_remove_value_from_product_attributes_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('product_attributes', function (Blueprint $table) {
-            $table->dropColumn('value');
+            if (Schema::hasColumn('product_attributes', 'value')) {
+                $table->dropColumn('value');
+            }
         });
     }
